+++ pybiscus/commands/app_local.py
@@ -7,8 +7,8 @@ from lightning.pytorch.callbacks import RichModelSummary, RichProgressBar
 from lightning.pytorch.callbacks.progress.rich_progress import RichProgressBarTheme
 from omegaconf import OmegaConf
 
-from src.console import console
-from src.ml.registry import datamodule_registry, model_registry
+from pybiscus.console import console
+from pybiscus.ml.registry import datamodule_registry, model_registry
 
 app = typer.Typer(pretty_exceptions_show_locals=False, rich_markup_mode="rich")
